Change Target.parse_makefile_child() to returning bool instead of raising exceptions
Now when Target.parse_makefile_child() encounters an element it doesn’t
recognize, it will return False instead of raising an exception.

This allows to chain calls to multiple implementations of
parse_makefile_child() as it’s needed by classes that derive from more
than one Target-derived class, as it’s the case for
ExecutableUnitTestTarget.